+++ crates/cubecl-matmul/src/components/global/read/strategy/async_barrier.rs
@@ -14,7 +14,7 @@ pub trait BarrierKind {
     fn level() -> BarrierLevel;
 }
 
-/// Asynchronous barrier for TMA loads
+/// Asynchronous barrier for `async_memcpy`
 pub struct AsyncBarrier {}
 
 #[cube]
@@ -29,6 +29,26 @@ impl SyncStrategy for AsyncBarrier {
         barrier: &mut Self::Barrier,
         #[comptime] _config: G,
     ) {
+        barrier.arrive_and_wait();
+    }
+}
+
+/// Asynchronous barrier for `async_copy`
+pub struct AsyncCopy {}
+
+#[cube]
+impl SyncStrategy for AsyncCopy {
+    type Barrier = Barrier;
+
+    fn create_barrier() -> Self::Barrier {
+        Barrier::new(BarrierLevel::cube_full(UNIT_POS == 0))
+    }
+
+    fn sync<MP: MatmulPrecision, G: GlobalConfig>(
+        barrier: &mut Self::Barrier,
+        #[comptime] _config: G,
+    ) {
+        barrier.commit_copy_async();
         barrier.arrive_and_wait();
     }
 }
